🔩 refactor(view): three.js logic and fix models not showing up on redirect
