+++ Views/css/style.css
@@ -303,6 +303,7 @@ figure {
 }
 .signin-figure { 
   margin-top: 0px;
+  margin-left: 50px;
 }
 
 .form-submit {
